# docs/agent/agent-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fractal Agent Gateway",
    "version": "1.0.0",
    "description": "Versioned, scoped surface for AI agents. Distinct from the human web API (/api/...). All routes (except /health) require a Bearer agent token. See docs/agent/AI_INTEGRATION_DESIGN.md."
  },
  "servers": [
    { "url": "http://localhost:8888" }
  ],
  "components": {
    "securitySchemes": {
      "AgentToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "qd_agent_xxx"
      },
      "HumanJWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "BacktestRequest": {
        "type": "object",
        "required": ["code","market","symbol","timeframe","start_date","end_date"],
        "properties": {
          "code":             { "type": "string", "description": "Indicator code (Python)" },
          "market":           { "type": "string", "example": "Crypto" },
          "symbol":           { "type": "string", "example": "BTC/USDT" },
          "timeframe":        { "type": "string", "example": "1D" },
          "start_date":       { "type": "string", "example": "2024-01-01" },
          "end_date":         { "type": "string", "example": "2024-12-31" },
          "initial_capital":  { "type": "number", "default": 10000 },
          "commission":       { "type": "number", "default": 0.001 },
          "slippage":         { "type": "number", "nullable": true },
          "leverage":         { "type": "integer", "default": 1 },
          "trade_direction":  { "type": "string", "default": "long" },
          "strictMode":       { "type": "boolean", "default": true, "description": "When true, signals execute on next bar open (live-aligned). When false, uses aggressive MTF/bar path." },
          "strict_mode":      { "type": "boolean", "description": "Alias for strictMode" },
          "strategy_config":  { "type": "object" },
          "indicator_params": { "type": "object" }
        }
      },
      "IndicatorValidateRequest": {
        "type": "object",
        "required": ["code"],
        "properties": {
          "code":             { "type": "string" },
          "indicator_params": { "type": "object" }
        }
      },
      "IndicatorSaveRequest": {
        "type": "object",
        "required": ["code"],
        "properties": {
          "code":             { "type": "string", "description": "Python indicator source (max 512 KiB)" },
          "name":             { "type": "string" },
          "description":      { "type": "string" },
          "indicator_id":     { "type": "integer", "description": "Update existing row when set" },
          "validate":         { "type": "boolean", "default": true }
        }
      },
...
    }
  },
  "security": [{ "AgentToken": [] }],
  "paths": {
...
    "/api/agent/v1/markets/{market}/symbols": {
      "get": {
        "summary": "Search symbols within a market",
        "x-scope-class": "R",
        "parameters": [
          { "in": "path",  "name": "market",  "required": true, "schema": { "type": "string" } },
          { "in": "query", "name": "keyword", "required": false, "schema": { "type": "string" } },
          { "in": "query", "name": "limit",   "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 100 } }
        ],
        "responses": { "200": { "description": "OK" }, "403": { "description": "Market not in token allowlist" } }
      }
    },
    "/api/agent/v1/klines": {
      "get": {
        "summary": "OHLCV bars",
        "x-scope-class": "R",
        "parameters": [
          { "in": "query", "name": "market",      "required": true,  "schema": { "type": "string" } },
          { "in": "query", "name": "symbol",      "required": true,  "schema": { "type": "string" } },
          { "in": "query", "name": "timeframe",   "required": false, "schema": { "type": "string", "default": "1D" } },
          { "in": "query", "name": "limit",       "required": false, "schema": { "type": "integer", "default": 300, "maximum": 2000 } },
          { "in": "query", "name": "before_time", "required": false, "schema": { "type": "integer", "description": "unix seconds" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/agent/v1/price": {
      "get": {
        "summary": "Latest price",
        "x-scope-class": "R",
        "parameters": [
          { "in": "query", "name": "market", "required": true, "schema": { "type": "string" } },
          { "in": "query", "name": "symbol", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/agent/v1/strategies": {
      "get":  {
        "summary": "List strategies (tenant-scoped)",
        "x-scope-class": "R",
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "summary": "Create a strategy (status defaults to 'stopped')",
        "x-scope-class": "W",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "Created" } }
      }
    },
    "/api/agent/v1/strategies/{strategy_id}": {
      "get": {
        "summary": "Get a strategy",
        "description": "Returns full strategy row for the tenant. Credential-like fields in exchange_config / notification_config are redacted (***) before the response leaves the Gateway.",
        "x-scope-class": "R",
        "parameters": [{ "in": "path", "name": "strategy_id", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not found" } }
      },
      "patch": {
        "summary": "Update a strategy (status='running' requires T scope)",
        "x-scope-class": "W",
        "parameters": [{ "in": "path", "name": "strategy_id", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/agent/v1/backtests": {
      "post": {
        "summary": "Submit a backtest job",
        "x-scope-class": "B",
        "parameters": [
          { "in": "header", "name": "Idempotency-Key", "required": false, "schema": { "type": "string" } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BacktestRequest" } } } },
        "responses": { "202": { "description": "Queued" } }
      }
    },
...
    "/api/agent/v1/indicators": {
      "get": {
        "summary": "List tenant indicators (compact; no code bodies)",
        "x-scope-class": "R",
        "parameters": [
          { "in": "query", "name": "limit", "required": false, "schema": { "type": "integer", "default": 50, "maximum": 200 } }
        ],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "summary": "Save indicator into private library (qd_indicator_codes)",
        "x-scope-class": "W",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IndicatorSaveRequest" } } }
        },
        "responses": { "200": { "description": "Saved" }, "400": { "description": "Validation failed or code too large" } }
      }
    },
    "/api/agent/v1/indicators/validate": {
      "post": {
        "summary": "Sandbox-validate indicator code without persisting",
        "x-scope-class": "R",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IndicatorValidateRequest" } } }
        },
        "responses": { "200": { "description": "Validation result" } }
      }
    },
...
    "/api/agent/v1/indicators/{indicator_id}": {
      "get": {
        "summary": "Fetch one indicator including Python source",
        "x-scope-class": "R",
        "parameters": [{ "in": "path", "name": "indicator_id", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not found" } }
      }
    },
...
    "/api/agent/v1/experiments/pipeline": {
      "post": {
        "summary": "Submit grid pipeline job",
        "x-scope-class": "B",
        "parameters": [{ "in": "header", "name": "Idempotency-Key", "required": false, "schema": { "type": "string" } }],
        "responses": { "202": { "description": "Queued" } }
      }
    },
...
    "/api/agent/v1/jobs": {
      "get": {
        "summary": "List recent jobs",
        "x-scope-class": "R",
        "parameters": [
          { "in": "query", "name": "kind",  "required": false, "schema": { "type": "string" } },
          { "in": "query", "name": "limit", "required": false, "schema": { "type": "integer", "default": 50 } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/agent/v1/jobs/{job_id}": {
      "get": {
        "summary": "Get a single job",
        "x-scope-class": "R",
        "parameters": [{ "in": "path", "name": "job_id", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not found" } }
      }
    },
    "/api/agent/v1/jobs/{job_id}/stream": {
      "get": {
        "summary": "Stream job progress as Server-Sent Events",
        "description": "Long-lived SSE stream. Frame types: `snapshot` (current job row), `progress` (each on_progress emission, with monotonic `seq`), `ping` (~15s keepalive), `result` (final). Resume after disconnect via `?since=<seq>` or `Last-Event-ID` header. If the job has already terminated when the client connects, `snapshot` and `result` are emitted immediately and the stream closes.",
        "x-scope-class": "R",
        "parameters": [
          { "in": "path", "name": "job_id", "required": true, "schema": { "type": "string" } },
          { "in": "query", "name": "since", "required": false, "schema": { "type": "integer", "minimum": 0 }, "description": "Resume from this sequence number (exclusive)." },
          { "in": "header", "name": "Last-Event-ID", "required": false, "schema": { "type": "string" }, "description": "Standard SSE resume header; alternative to `since`." }
        ],
        "responses": {
          "200": {
            "description": "SSE stream",
            "content": { "text/event-stream": { "schema": { "type": "string" } } }
          },
          "404": { "description": "Job not found in this tenant" }
        }
      }
    },
...
  }
}